+++ packages/isaacscript-common/src/enums/StageTravelState.ts
@@ -1,0 +1,6 @@
+export enum StageTravelState {
+  DISABLED,
+  FADING_TO_BLACK,
+  GOING_TO_NEW_FLOOR,
+  FADING_IN,
+}