+++ packages/isaacscript-common/src/features/customTrapdoor/constants.ts
@@ -1,0 +1,7 @@
+export const STAGE_TRAVEL_FEATURE_NAME = "stageTravel";
+
+/** This also applies to crawl spaces. The value was determined through trial and error. */
+export const TRAPDOOR_OPEN_DISTANCE = 60;
+
+export const TRAPDOOR_OPEN_DISTANCE_AFTER_BOSS = TRAPDOOR_OPEN_DISTANCE * 2.5;
+export const TRAPDOOR_BOSS_REACTION_FRAMES = 30;