+++ packages/isaacscript-common/src/functions/stage.ts
@@ -3,6 +3,66 @@ import { game } from "../cachedClasses";
 import { STAGE_TYPE_TO_LETTER } from "../objects/stageTypeToLetter";
 
 /**
+ * Helper function that calculates what the stage type should be for the provided stage. This
+ * emulates what the game's internal code does.
+ */
+export function calculateStageType(stage: LevelStage): StageType {
+  // The following is the game's internal code to determine the floor type. (This came directly from
+  // Spider.)
+
+  /*
+    u32 Seed = g_Game->GetSeeds().GetStageSeed(NextStage);
+    if (!g_Game->IsGreedMode()) {
+      StageType = ((Seed % 2) == 0 && (
+        ((NextStage == STAGE1_1 || NextStage == STAGE1_2) && gd.Unlocked(ACHIEVEMENT_CELLAR)) ||
+        ((NextStage == STAGE2_1 || NextStage == STAGE2_2) && gd.Unlocked(ACHIEVEMENT_CATACOMBS)) ||
+        ((NextStage == STAGE3_1 || NextStage == STAGE3_2) && gd.Unlocked(ACHIEVEMENT_NECROPOLIS)) ||
+        ((NextStage == STAGE4_1 || NextStage == STAGE4_2)))
+      ) ? STAGE_TYPE_WOTL : STAGE_TYPE_ORIGINAL;
+    if (Seed % 3 == 0 && NextStage < STAGE5)
+      StageType = STAGE_TYPE_AFTERBIRTH;
+  */
+
+  const seeds = game.GetSeeds();
+  const stageSeed = seeds.GetStageSeed(stage);
+
+  if (stageSeed % 2 === 0) {
+    return StageType.WRATH_OF_THE_LAMB;
+  }
+
+  if (stageSeed % 3 === 0) {
+    return StageType.AFTERBIRTH;
+  }
+
+  return StageType.ORIGINAL;
+}
+
+/**
+ * Helper function that calculates what the Repentance stage type should be for the provided stage.
+ * This emulates what the game's internal code does.
+ */
+export function calculateStageTypeRepentance(stage: LevelStage): StageType {
+  // There is no alternate floor for Corpse.
+  if (stage === LevelStage.WOMB_1 || stage === LevelStage.WOMB_2) {
+    return StageType.REPENTANCE;
+  }
+
+  // This algorithm is from Kilburn. We add one because the alt path is offset by 1 relative to the
+  // normal path.
+  const seeds = game.GetSeeds();
+  const adjustedStage = ((stage as int) + 1) as LevelStage;
+  const stageSeed = seeds.GetStageSeed(adjustedStage);
+
+  // Kilburn does not know why he divided the stage seed by 2 first.
+  const halfStageSeed = Math.floor(stageSeed / 2);
+  if (halfStageSeed % 2 === 0) {
+    return StageType.REPENTANCE_B;
+  }
+
+  return StageType.REPENTANCE;
+}
+
+/**
  * Helper function to account for Repentance floors being offset by 1. For example, Downpour 2 is
  * the third level of the run, but the game considers it to have a stage of 2. This function will
  * consider Downpour 2 to have a stage of 3.