+++ packages/isaacscript-common/src/interfaces/StageTravelEntityDescription.ts
@@ -1,0 +1,6 @@
+import { StageTravelEntityState } from "../enums/StageTravelEntityState";
+
+export interface StageTravelEntityDescription {
+  initial: boolean;
+  state: StageTravelEntityState;
+}